fix: resolve all clippy warnings across workspace
Eliminated ~90 warnings across 8 crates:
- nsl-errors/nsl-semantic/nsl-parser: redundant closures, collapsible ifs,
  needless return, new_without_default, while_let_loop, match→if_let
- nsl-ast: box large TrainSection::Stmt enum variant
- nsl-runtime/cuda: add #![allow(dead_code)] for PTX kernel constants
  loaded at runtime via CUDA driver API string lookup
- nsl-runtime/tensor,cpu,autodiff: needless_range_loop (use enumerate),
  unnecessary_cast, manual_slice_size_calculation, redundant_field_names,
  too_many_arguments (#[allow]), unused variable
- nsl-runtime/dataloader,memory,packing: dead_code fields (future API),
  not_unsafe_ptr_arg_deref (#[allow] on extern C FFI functions)
- nsl-runtime/tokenizer: box large HuggingFace variant
- nsl-runtime/quantize: div_ceil, is_multiple_of, needless_range_loop
- nsl-runtime/huggingface,sampling,file_io: cmp_null, redundant_pattern_matching
- nsl-codegen: new_without_default (Default impls), type_complexity (#[allow]),
  needless_borrow, unnecessary_cast, useless_format, too_many_arguments,
  map_or→is_some_and, len_zero
- nsl-cli: dead_code fields/functions (#[allow]), manual_pattern_char_comparison,
  type_complexity (#[allow]), ptr_arg (&Path), unnecessary_unwrap (if let)

+++ crates/nsl-codegen/src/expr.rs
@@ -1243,7 +1243,7 @@ impl Compiler<'_> {
             let dim_val = if args.len() > 2 {
                 self.compile_expr(builder, state, &args[2].value)?
             } else {
-                builder.ins().iconst(cl_types::I64, -1i64 as i64)
+                builder.ins().iconst(cl_types::I64, -1_i64)
             };
             return self.compile_call_by_name(builder, "nsl_tensor_topk", &[tensor_val, k_val, dim_val]);
         }
@@ -1261,7 +1261,7 @@ impl Compiler<'_> {
             let dim_val = if args.len() > 1 {
                 self.compile_expr(builder, state, &args[1].value)?
             } else {
-                builder.ins().iconst(cl_types::I64, -1i64 as i64)
+                builder.ins().iconst(cl_types::I64, -1_i64)
             };
             return self.compile_call_by_name(builder, "nsl_tensor_argmax", &[tensor_val, dim_val]);
         }
@@ -2666,6 +2666,7 @@ impl Compiler<'_> {
 
     // ── Tensor operations ──────────────────────────────────────────
 
+    #[allow(clippy::too_many_arguments)]
     fn compile_tensor_binary_op(
         &mut self,
         builder: &mut FunctionBuilder,
